+++ src/gas_management/__init__.py
@@ -8,5 +8,5 @@ from __future__ import annotations
 
 __all__ = ["__version__", "__author__"]
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"
 __author__ = "Aashish Bharti"